+++ packages/intents-sdk/src/bridges/direct-bridge/direct-bridge.ts
@@ -27,6 +27,8 @@ import {
 	createWithdrawIntentPrimitive,
 	withdrawalParamsInvariant,
 } from "./direct-bridge-utils";
+import { UnsupportedAssetIdError } from "../../classes/errors";
+import { parseDefuseAssetId } from "../../lib/parse-defuse-asset-id";
 
 export class DirectBridge implements Bridge {
 	protected env: NearIntentsEnv;
@@ -44,22 +46,27 @@ export class DirectBridge implements Bridge {
 		return routeConfig.route === RouteEnum.NearWithdrawal;
 	}
 
-	supports(params: Pick<WithdrawalParams, "assetId" | "routeConfig">): boolean {
-		let result = true;
-
-		if ("routeConfig" in params && params.routeConfig != null) {
-			result &&= this.is(params.routeConfig);
-		}
-
-		try {
-			return result && this.parseAssetId(params.assetId) != null;
-		} catch {
+	async supports(
+		params: Pick<WithdrawalParams, "assetId" | "routeConfig">,
+	): Promise<boolean> {
+		if (params.routeConfig != null && !this.is(params.routeConfig)) {
 			return false;
 		}
+
+		const assetInfo = this.parseAssetId(params.assetId);
+		const isValid = assetInfo != null;
+
+		if (!isValid && params.routeConfig != null) {
+			throw new UnsupportedAssetIdError(
+				params.assetId,
+				"`assetId` does not match `routeConfig`.",
+			);
+		}
+		return isValid;
 	}
 
 	parseAssetId(assetId: string): ParsedAssetInfo | null {
-		const parsed = utils.parseDefuseAssetId(assetId);
+		const parsed = parseDefuseAssetId(assetId);
 
 		if (parsed.standard === "nep141") {
 			return Object.assign(parsed, {