+++ packages/intents-sdk/src/bridges/hot-bridge/hot-bridge.test.ts
@@ -1,0 +1,98 @@
+import { describe, expect, it } from "vitest";
+import { HotBridge } from "./hot-bridge";
+import { UnsupportedAssetIdError } from "../../classes/errors";
+import hotOmniSdk from "@hot-labs/omni-sdk";
+import { createHotBridgeRoute } from "../../lib/route-config-factory";
+import { Chains } from "../../lib/caip2";
+
+describe("HotBridge", () => {
+	describe("supports()", () => {
+		it.each([
+			"nep245:v2_1.omni.hot.tg:56_11111111111111111111",
+			// todo: add more test cases for each supported chain
+		])("supports `v2_1.omni.hot.tg` tokens", async (tokenId) => {
+			const bridge = new HotBridge({
+				env: "production",
+				hotSdk: new hotOmniSdk.HotBridge({
+					async executeNearTransaction() {
+						throw new Error("not implemented");
+					},
+				}),
+			});
+
+			await expect(bridge.supports({ assetId: tokenId })).resolves.toBe(true);
+			await expect(
+				bridge.supports({
+					assetId: tokenId,
+					routeConfig: createHotBridgeRoute(Chains.TON),
+				}),
+			).resolves.toBe(true);
+		});
+
+		it.each([
+			"nep141:btc.omft.near",
+			"nep245:v3_1.omni.hot.tg:56_11111111111111111111",
+		])("doesn't support `v2_1.omni.hot.tg` tokens", async (tokenId) => {
+			const bridge = new HotBridge({
+				env: "production",
+				hotSdk: new hotOmniSdk.HotBridge({
+					async executeNearTransaction() {
+						throw new Error("not implemented");
+					},
+				}),
+			});
+			await expect(bridge.supports({ assetId: tokenId })).resolves.toBe(false);
+		});
+
+		it.each([
+			"nep141:v2_1.omni.hot.tg:56_11111111111111111111",
+			"nep245:v2_1.omni.hot.tg",
+			"nep245:v2_1.omni.hot.tg:FOOBAR",
+		])(
+			"throws UnsupportedAssetIdError if misspelled HOT token",
+			async (assetId) => {
+				const bridge = new HotBridge({
+					env: "production",
+					hotSdk: new hotOmniSdk.HotBridge({
+						async executeNearTransaction() {
+							throw new Error("not implemented");
+						},
+					}),
+				});
+
+				await expect(bridge.supports({ assetId: assetId })).rejects.toThrow(
+					UnsupportedAssetIdError,
+				);
+
+				// It throws even if routeConfig is provided.
+				await expect(
+					bridge.supports({
+						assetId: assetId,
+						routeConfig: createHotBridgeRoute(Chains.TON),
+					}),
+				).rejects.toThrow(UnsupportedAssetIdError);
+			},
+		);
+
+		it.each(["invalid_string", "nep141:wrap.near", "nep141:btc.omft.near"])(
+			"throws UnsupportedAssetIdError if routeConfig passed, but assetId is not HOT token",
+			async (assetId) => {
+				const bridge = new HotBridge({
+					env: "production",
+					hotSdk: new hotOmniSdk.HotBridge({
+						async executeNearTransaction() {
+							throw new Error("not implemented");
+						},
+					}),
+				});
+
+				await expect(
+					bridge.supports({
+						assetId,
+						routeConfig: createHotBridgeRoute(Chains.TON),
+					}),
+				).rejects.toThrow(UnsupportedAssetIdError);
+			},
+		);
+	});
+});