+++ packages/intents-sdk/src/sdk.ts
@@ -118,6 +118,10 @@ export class IntentsSDK implements IIntentsSDK {
 					},
 				}),
 			}),
+			// new OmniBridge({
+			// 	env: this.env,
+			// 	nearProvider,
+			// }),
 			new DirectBridge({
 				env: this.env,
 				nearProvider,
@@ -140,7 +144,7 @@ export class IntentsSDK implements IIntentsSDK {
 		logger?: ILogger;
 	}): Promise<IntentPrimitive[]> {
 		for (const bridge of this.bridges) {
-			if (bridge.supports(args.withdrawalParams)) {
+			if (await bridge.supports(args.withdrawalParams)) {
 				const actualAmount = args.withdrawalParams.feeInclusive
 					? args.withdrawalParams.amount - args.feeEstimation.amount
 					: args.withdrawalParams.amount;
@@ -149,6 +153,8 @@ export class IntentsSDK implements IIntentsSDK {
 					assetId: args.withdrawalParams.assetId,
 					amount: actualAmount,
 					destinationAddress: args.withdrawalParams.destinationAddress,
+					feeEstimation: args.feeEstimation,
+					routeConfig: args.withdrawalParams.routeConfig,
 					logger: args.logger,
 				});
 
@@ -208,7 +214,7 @@ export class IntentsSDK implements IIntentsSDK {
 		logger?: ILogger;
 	}): Promise<FeeEstimation> {
 		for (const bridge of this.bridges) {
-			if (bridge.supports(args.withdrawalParams)) {
+			if (await bridge.supports(args.withdrawalParams)) {
 				const fee = await bridge.estimateWithdrawalFee({
 					withdrawalParams: args.withdrawalParams,
 					quoteOptions: args.quoteOptions,