+++ packages/internal-utils/src/utils/prepareBroadcastRequest.ts
@@ -66,13 +66,29 @@ export function prepareSwapSignedData(
 			};
 		}
 
-		case "STELLAR": {
+		case "STELLAR_RAW": {
 			assert(
 				userInfo.userChainType === "stellar",
 				"User chain and signature chain must match",
 			);
 			return {
 				standard: "raw_ed25519",
+				payload: signature.signedData.message,
+				// We should encode the Stellar address to base58
+				public_key: `ed25519:${base58.encode(
+					stellarAddressToBytes(userInfo.userAddress),
+				)}`,
+				signature: transformED25519Signature(signature.signatureData),
+			};
+		}
+
+		case "STELLAR_SEP53": {
+			assert(
+				userInfo.userChainType === "stellar",
+				"User chain and signature chain must match",
+			);
+			return {
+				standard: "sep53",
 				payload: signature.signedData.message,
 				// We should encode the Stellar address to base58
 				public_key: `ed25519:${base58.encode(