+++ packages/internal-utils/src/utils/promise/withTimeout.ts
@@ -3,7 +3,9 @@ import { BaseError } from "../../errors/base";
 export function withTimeout<data>(
 	fn: ({
 		signal,
-	}: { signal: AbortController["signal"] | null }) => Promise<data>,
+	}: {
+		signal: AbortController["signal"] | null;
+	}) => Promise<data>,
 	{
 		errorInstance = new Error("timed out"),
 		timeout,